Duty-Chart-Management-Backend-main/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pillow==11.3.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.2
setuptools==80.9.0
sqlparse==0.5.3
uritemplate==4.2.0
wheel==0.45.1
pandas>=2.2.2,<2.3
openpyxl>=3.1.2,<3.2
